OpenRCT2 Initial Commit
A very initial commit for OpenRCT2 GenericTemplate. With some major issues (tested on Windows):
- When stopping Server, the server keeps running. 
- Console seems to not be operational as of yet.

Linux:
I'll need to test on other distrubutions then 20.04, but it seems that 20.04 LTS is not supported at this moment without downloading and running using ppa/apt due to dependency errors.

There will also need to be more done on the configmanifest.